+++ cwolfmap/audio.c
@@ -53,7 +53,7 @@ int CWAudioLoadAudioT(CWAudio *audio, const char *path)
 		fprintf(stderr, "Failed to read %s", path);
 		goto bail;
 	}
-	const int len = audio->head.offsets[audio->head.nOffsets - 1];
+	const uint32_t len = audio->head.offsets[audio->head.nOffsets - 1];
 	audio->data = malloc(len);
 	if (fread(audio->data, 1, len, f) != len)
 	{
@@ -79,7 +79,7 @@ void CWAudioFree(CWAudio *audio)
 }
 
 int CWAudioGetAdlibSound(
-	const CWAudio *audio, const int i, const char **data, int *len)
+	const CWAudio *audio, const int i, const char **data, size_t *len)
 {
 	int err = 0;
 	const int off = audio->head.offsets[i + STARTADLIBSOUNDS];
@@ -97,7 +97,7 @@ bail:
 }
 
 int CWAudioGetMusic(
-	const CWAudio *audio, const int i, const char **data, int *len)
+	const CWAudio *audio, const int i, const char **data, size_t *len)
 {
 	int err = 0;
 	const int off = audio->head.offsets[i + STARTMUSIC];
